Move out of settings when face is changed
While using the Countdown Face, I noticed that I would sometimes change the amount of time to countdown from instead of starting the timer. This occurred despite the fact that the numbers were not blinking.

I believe this is due to leaving the settings mode and entering another watch face. The countdown face remains in settings mode, but since the movement tick frequency reverts back to 1, the blink no longer occurs.

This commit moves into waiting mode if the Countdown Face is in settings mode and the user moves to the next face.

The order of operations to reproduce this on the previous versions is as follows.

Move to Countdown Face -> Enter Settings Mode -> Move out of Countdown Face -> Move back to Countdown Face -> Attempt to start the timer with Alarm button, but the timer amount increases instead.

I tested this commit and it does fix the issue for me.