Optimization: compare raw ADC battery readings
This is a silly optimization to make. It saves about four instructions
once a day when the battery is checked. There is a tiny bit of extra
hidden savings: those instructions happen while the ADC is enabled.
I'm sure it saves whole picojoules!

Note the operator change from less than to less than or equal. This is
due to a rounding difference when the raw ADC value is exactly equal to
the threshold value. I didn't work this through to a definitive answer
for where the rounding is happening. After the operator change, the code
displays equivalent behavior.

See this godbolt for more details https://godbolt.org/z/9W7vha5P6