Implementation of an animated clock face.
This watch face is an extension of the simple watch face and contains exactly the same functions. However, depending on a configurable interval, the watch face displays short animations to amaze your viewers and demonstrate the capabilities of the sensor watch.

Currently, the antimated watch face supports two animations that are displayed at an interval of 1, 5, 10, 15, 30 or 60 minutes. The animations are:

1. **scroll clock**. With this animation the current time is shown as text on the display, such as "ten twenty-two" for "10:22". Initially, I experimented with longer text but found it difficult to read due to the limitations of the segment display.

2. **wyoscan clock**. This animation uses the amazing wyoscan clock by Willy Hardy. The animation runs three times, where it first covers the displayed time, then moves over it, and finally restores it.

A *long press* on the *light button* activates the settings of this clock face. Currently, you can only select the interval for the animations or disable them. I would have liked to use a diferent button here, but the others are already occupied by the simple clock.

I tried to construct the animated clock face in a generic way. The main structure contains a general-purpose buffer that can be used by different animations to store temporary data. For example, the buffer contains the scrolling text for the scroll clock and the segment configuration of the wyoscan clock.