i18n: Add compile-time language setting.
This currently supports Spanish and German (with English being the
default), and simply translates the weekdays, although more translations
should be added.

German translations are from @Kaoala on GH Discussions.
Spanish translations are from ltzamatema on Discord.

DO is a little hard to read (it looks a little like OO if you're not
expecting it). Possibly dO would look better, but I've chosen to keep
everything in caps for now. If someone who uses the watch in another
language would like to change it, feel free.