Enable LTO.
This saves around 4% of space on the default build, and likely reduces
power consumption as well, although I'm not sure by how much. Numbers
from TI [1] suggest that it might be a ~2-3% speed improvement (they
measure 3.8% using -O3, but we're using -Os so we should expect slightly
less), which would probably translate to at least a few days of extra
battery life.

The change to the .vectors section is from a blog post [2] that did a
excellent job explaining the exact problem I had.

[1]: https://software-dl.ti.com/codegen/docs/articles/cgt_Build_a_Leaner_and_Faster_TI_Arm_Application_with_Link_Time_Optimization.html
[2]: https://m0agx.eu/fixing-cortex-m-startup-code-for-link-time-optimization.html