WIP switching RTC to count32 (vs clock)
Clock is nice, in that it:
- gives us year/month/day/hours/mins/secs straight up
- allows one to set a number of alarms

However, it doesn't support more than 1 sec granularity.
This makes things like a stopwatch (or time systems that don't use seconds) annoying,
and also makes it harder for us to set it appropriately (see hackwatch).

If we use COUNT32 instead, we get up to 1024Hz granularity (with the 1024Hz oscillator configured).
This commit has it set to 256Hz.

Downsides:
- extra code complexity
- we have to maintain an offset (and know about the TZ at a lower level if we want compatibility)
- we lose the in-built periodic alarms. At the moment, we're only using these for the 1minute wake from sleep
  operation, which could be replaced by a COMP check (TODO!)

TODO:
- simulator support
- comments on new public functions
- public function to get just the sub-second count (maybe?)
  or create a non-compatible watch_get_date_time function
  which returns all the relevant info (and the real year?)
- add COMP check to replicate wake from sleep behaviour
- power consumption implications (if none, switch to 1024Hz counter?)