Improve the credential management.
When storing many TOTP credentials, it could get bothersome to keep
all the arrays synchronized. Now there are the `credentials` array
which holds most of the parameters and the `keys` array which is a
concession to not have the storage waste or the malloc that would be
required when storing the key in the `totp_parameters_t` struct along
with everything else.
Additionally there is a helper program found at `utils/totp_face_helper.py`
which properly transforms the key from its base32 representation into
the form expected by the TOTP watch face.