movement: conditionally compile 24h mode forcer
Define a new static function that forces the movement to 24 hour mode
if configured to do so at build time.

By default, the watch will not be configured to do it
which means the code will not be compiled into the function.
The compiler's dead code elimination should delete all references to it.