preferences: conditionally compile page skipper
Define a new static function that skips the clock mode page
in the preferences watch face if configured to force 24h display.

By default, the watch will not be configured in forced 24h mode
which means the code will not be compiled into the function.
The compiler's dead code elimination should delete all references to it.