faces/clock: add CLOCK_FACE_SIGNAL_INTERVAL value
One of the most iconic features of the Casio F91W watch is the hourly
signal which creates an audible beep every 60 minutes allowing the
wearer to keep track of time without looking at their watch.

Building upon this concept there have been references to this specific
feature in popular culture such as the TV Show 'Mr Robot'.

> Each beep indicates one minute of my time that has passed. I have
> allotted you no more than three minutes.
-- Whiterose

However, the watch seen in the show doesn't actually exist as the signal
interval on Casio watches is not configurable. This commit allows the
user to define a compile time constant making the beep operate at a more
frequent interval than the hourly default.

In the future this should be configurable at runtime.

Signed-off-by: J. Dekker <jdek@itanimul.li>