Revert merge of PR #437 - debouncing logic
This reverts commit a715265af61ecab41eaefe0c5a7e7c806d0db41b, reversing
changes made to 9c093f954084e61614cba1c259e4581a87420813.

Insidious issues were found in the course of long term testing
by the community, and further reviews of the code were not enough
to pinpoint the issue and fix it. So for now the appropriate action
is to revert these changes while development continues, and possibly
merge them back in once they have been stabilized.

Tested-on-hardware-by: David Volovskiy <devolov@gmail.com>
Tested-on-hardware-by: CarpeNoctem <cryptomax@pm.me>
Tested-on-hardware-by: Krzysztof Gałka <@kshysztof@Discord>